+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added list of supported sharing services to READMEs.
 * Added `{url}` placeholder to `data-mail-body` option. (stephankellermayr)
 * Fixed Twitter popup opening twice when a tweet is embedded on a page. (Nebel54)
+* Improved service initialization code. [[#188](https://github.com/heiseonline/shariff/issues/188)]
 
 ## 1.25.2 2017-06-13